POST /yesterday
Load notices from MySQL filtered by yesterday's date (21.07.16)

Result:

[
    {
        "id": 13734,
        "title": "2021 건물에너지관리시스템(BEMS) 교육 프로그램 신청접수(~7/21)",
        "date": "21.07.16",
        "link": "https://www.ajou.ac.kr/kr/ajou/notice.do?mode=view&articleNo=111747&article.offset=0&articleLimit=15",
        "writer": "공학교육혁신센터"
    }
]

+++ src/db/models.rs
@@ -1,11 +1,13 @@
 #![allow(proc_macro_derive_resolution_fallback)]
 
+use crate::db::schema::ajou_notices;
 use crate::db::schema::ajou_sched;
 // use diesel::prelude::*;
 
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Queryable, AsChangeset, Serialize, Deserialize, Debug, Clone)]
+#[table_name = "ajou_notices"]
 pub struct Notice {
-    pub id: u64,
+    pub id: i32,
     pub title: String,
     pub date: String,
     pub link: String,
